adding sw
