+++ package.json
@@ -5,7 +5,11 @@
   "homepage": "https://variadicjs.github.io/",
   "dependencies": {
     "babel-preset-es2015": "^6.24.1",
+    "body-parser": "^1.18.2",
+    "cra-append-sw": "^2.2.0",
+    "express": "^4.16.3",
     "gh-pages": "^1.1.0",
+    "path": "^0.12.7",
     "react": "^16.2.0",
     "react-dom": "^16.2.0",
     "react-materialize": "^1.1.2",
@@ -28,8 +32,8 @@
   "scripts": {
     "predeploy": "npm run build",
     "deploy": "gh-pages -b master -d build",
-    "start": "react-scripts start",
-    "build": "react-scripts build",
+    "start": "react-scripts start && cra-append-sw --mode dev ./sw.js",
+    "build": "react-scripts build && cra-append-sw --mode build --skip-compile ./sw.js", 
     "test": "jest --watch",
     "coverage": "jest --coverage",
     "eject": "react-scripts eject"
