+++ src/index.js
@@ -1,8 +1,8 @@
 import React from 'react';
 import ReactDOM from 'react-dom';
 import App from './containers/App';
-// import registerServiceWorker from './registerServiceWorker';
+import registerCustomSW from './registerCustomSW';
 
 ReactDOM.render(<App />, document.getElementById('root'));
 
-// registerServiceWorker();
+registerCustomSW();
